visual_search_system/requirements.txt
Flask
matplotlib
numpy
pandas
python-dateutil
scikit-learn
torch
torchaudio
torchvision